template components all imported from index
