+++ app/components/templates/checkbox-field.tsx
@@ -1,0 +1,67 @@
+import { useInputControl } from '@conform-to/react'
+import { useId } from 'react'
+import { Checkbox, type CheckboxProps } from '../ui'
+import { ErrorList, type ListOfErrors } from '.'
+
+export function CheckboxField({
+	labelProps,
+	buttonProps,
+	errors,
+	className,
+}: {
+	labelProps: JSX.IntrinsicElements['label']
+	buttonProps: CheckboxProps & {
+		name: string
+		form: string
+		value?: string
+	}
+	errors?: ListOfErrors
+	className?: string
+}) {
+	const { key, defaultChecked, ...checkboxProps } = buttonProps
+	const fallbackId = useId()
+	const checkedValue = buttonProps.value ?? 'on'
+	const input = useInputControl({
+		key,
+		name: buttonProps.name,
+		formId: buttonProps.form,
+		initialValue: defaultChecked ? checkedValue : undefined,
+	})
+	const id = buttonProps.id ?? fallbackId
+	const errorId = errors?.length ? `${id}-error` : undefined
+
+	return (
+		<div className={className}>
+			<div className="flex gap-2">
+				<Checkbox
+					{...checkboxProps}
+					id={id}
+					aria-invalid={errorId ? true : undefined}
+					aria-describedby={errorId}
+					checked={input.value === checkedValue}
+					onCheckedChange={state => {
+						input.change(state.valueOf() ? checkedValue : '')
+						buttonProps.onCheckedChange?.(state)
+					}}
+					onFocus={event => {
+						input.focus()
+						buttonProps.onFocus?.(event)
+					}}
+					onBlur={event => {
+						input.blur()
+						buttonProps.onBlur?.(event)
+					}}
+					type="button"
+				/>
+				<label
+					htmlFor={id}
+					{...labelProps}
+					className="self-center text-body-xs text-muted-foreground"
+				/>
+			</div>
+			<div className="px-4 pb-3 pt-1">
+				{errorId ? <ErrorList id={errorId} errors={errors} /> : null}
+			</div>
+		</div>
+	)
+}
