+++ app/components/templates/errors.tsx
@@ -1,0 +1,21 @@
+export type ListOfErrors = Array<string | null | undefined> | null | undefined
+
+export function ErrorList({
+	id,
+	errors,
+}: {
+	errors?: ListOfErrors
+	id?: string
+}) {
+	const errorsToRender = errors?.filter(Boolean)
+	if (!errorsToRender?.length) return null
+	return (
+		<ul id={id} className="flex flex-col gap-1">
+			{errorsToRender.map(e => (
+				<li key={e} className="text-[10px] text-foreground-destructive">
+					{e}
+				</li>
+			))}
+		</ul>
+	)
+}
