+++ app/components/templates/progress-bar.tsx
@@ -2,7 +2,7 @@ import { useNavigation } from '@remix-run/react'
 import { useEffect, useRef, useState } from 'react'
 import { useSpinDelay } from 'spin-delay'
 import { cn } from '#app/utils/misc.tsx'
-import { Icon } from './ui/icon.tsx'
+import { Icon } from '../ui/icon.tsx'
 
 function EpicProgress() {
 	const transition = useNavigation()
