+++ app/components/templates/search-bar.tsx
@@ -1,10 +1,8 @@
 import { Form, useSearchParams, useSubmit } from '@remix-run/react'
 import { useId } from 'react'
 import { useDebounce, useIsPending } from '#app/utils/misc.tsx'
-import { StatusButton } from './templates/index.ts'
-import { Icon } from './ui/icon.tsx'
-import { Input } from './ui/input.tsx'
-import { Label } from './ui/label.tsx'
+import { Icon, Input, Label } from '../ui'
+import { StatusButton } from '.'
 
 export function SearchBar({
 	status,
