+++ app/routes/_auth+/onboarding_.$provider.tsx
@@ -22,9 +22,13 @@ import {
 } from '@remix-run/react'
 import { safeRedirect } from 'remix-utils/safe-redirect'
 import { z } from 'zod'
-import { CheckboxField, ErrorList, Field } from '#app/components/forms.tsx'
-import { Spacer } from '#app/components/spacer.tsx'
-import { StatusButton } from '#app/components/templates/index.ts'
+import {
+	CheckboxField,
+	ErrorList,
+	Field,
+	Spacer,
+	StatusButton,
+} from '#app/components/templates'
 import {
 	authenticator,
 	requireAnonymous,
