+++ app/routes/_auth+/signup.tsx
@@ -12,8 +12,7 @@ import { Form, useActionData, useSearchParams } from '@remix-run/react'
 import { HoneypotInputs } from 'remix-utils/honeypot/react'
 import { z } from 'zod'
 import { GeneralErrorBoundary } from '#app/components/error-boundary.tsx'
-import { ErrorList, Field } from '#app/components/forms.tsx'
-import { StatusButton } from '#app/components/templates/index.ts'
+import { ErrorList, Field, StatusButton } from '#app/components/templates'
 import {
 	ProviderConnectionForm,
 	providerNames,
