+++ app/routes/settings+/profile.password.tsx
@@ -9,9 +9,8 @@ import {
 } from '@remix-run/node'
 import { Form, Link, useActionData } from '@remix-run/react'
 import { z } from 'zod'
-import { ErrorList, Field } from '#app/components/forms.tsx'
-import { StatusButton } from '#app/components/templates/index.ts'
-import { Button, Icon } from '#app/components/ui/index.ts'
+import { ErrorList, Field, StatusButton } from '#app/components/templates'
+import { Button, Icon } from '#app/components/ui'
 import {
 	getPasswordHash,
 	requireUserId,
