+++ app/routes/settings+/profile.tsx
@@ -3,8 +3,8 @@ import { type SEOHandle } from '@nasa-gcn/remix-seo'
 import { json, type LoaderFunctionArgs } from '@remix-run/node'
 import { Link, Outlet, useMatches } from '@remix-run/react'
 import { z } from 'zod'
-import { Spacer } from '#app/components/spacer.tsx'
-import { Icon } from '#app/components/ui/index.ts'
+import { Spacer } from '#app/components/templates'
+import { Icon } from '#app/components/ui'
 import { requireUserId } from '#app/utils/auth.server.ts'
 import { prisma } from '#app/utils/db.server.ts'
 import { cn } from '#app/utils/misc.tsx'
