+++ app/routes/settings+/profile.two-factor.verify.tsx
@@ -15,9 +15,8 @@ import {
 } from '@remix-run/react'
 import * as QRCode from 'qrcode'
 import { z } from 'zod'
-import { ErrorList, Field } from '#app/components/forms.tsx'
-import { StatusButton } from '#app/components/templates/index.ts'
-import { Icon } from '#app/components/ui/index.ts'
+import { ErrorList, Field, StatusButton } from '#app/components/templates'
+import { Icon } from '#app/components/ui'
 import { isCodeValid } from '#app/routes/_auth+/verify.tsx'
 import { requireUserId } from '#app/utils/auth.server.ts'
 import { prisma } from '#app/utils/db.server.ts'
